Move block index to separate module
This is required in order to break dependencies loop for mempool which will need
block storage and state view which are now defined in Consensus module.
